+++ packages/instantsearch-ui-components/src/components/chat/Chat.tsx
@@ -7,7 +7,7 @@ import { createChatOverlayLayoutComponent } from './ChatOverlayLayout';
 import { createChatPromptComponent } from './ChatPrompt';
 import { createChatPromptSuggestionsComponent } from './ChatPromptSuggestions';
 
-import type { Renderer, ComponentProps } from '../../types';
+import type { Renderer, ComponentProps, Hooks } from '../../types';
 import type { ChatHeaderProps, ChatHeaderOwnProps } from './ChatHeader';
 import type { ChatMessagesProps } from './ChatMessages';
 import type { ChatPromptProps, ChatPromptOwnProps } from './ChatPrompt';
@@ -92,9 +92,17 @@ export type ChatProps = Omit<ComponentProps<'div'>, 'onError' | 'title'> & {
   layoutComponent?: (props: ChatLayoutOwnProps) => JSX.Element;
 };
 
-export function createChatComponent({ createElement, Fragment }: Renderer) {
+export function createChatComponent({
+  createElement,
+  Fragment,
+  memo,
+}: Renderer & Partial<Pick<Hooks, 'memo'>>) {
   const ChatHeader = createChatHeaderComponent({ createElement, Fragment });
-  const ChatMessages = createChatMessagesComponent({ createElement, Fragment });
+  const ChatMessages = createChatMessagesComponent({
+    createElement,
+    Fragment,
+    memo,
+  });
   const ChatPrompt = createChatPromptComponent({ createElement, Fragment });
   const ChatPromptSuggestions = createChatPromptSuggestionsComponent({
     createElement,