+++ packages/instantsearch-ui-components/src/components/chat/types.ts
@@ -449,11 +449,21 @@ export type AddToolResultWithOutput = (
   params: Pick<Parameters<AddToolResult>[0], 'output'>
 ) => ReturnType<AddToolResult>;
 
-export type SearchToolInput = {
+type SearchToolInputBase = {
   query: string;
   number_of_results?: number;
+};
+
+type DefaultSearchToolInput = SearchToolInputBase & {
   facet_filters?: string[][];
 };
+
+type McpSearchToolInput = SearchToolInputBase & {
+  facet_filters?: undefined;
+  [facetKey: `facet_${string}`]: string[] | undefined;
+};
+
+export type SearchToolInput = DefaultSearchToolInput | McpSearchToolInput;
 
 export type ApplyFiltersParams = {
   query?: string;
@@ -484,6 +494,7 @@ export type ChatLayoutOwnProps<
 
 export type ClientSideToolComponentProps = {
   message: ChatToolMessage;
+  messages?: ChatMessageBase[];
   indexUiState: object;
   setIndexUiState: (state: object) => void;
   onClose: () => void;
@@ -536,4 +547,3 @@ export type ChatEmptyProps = {
    */
   setInput?: (input: string) => void;
 };
-