+++ packages/react-instantsearch-nextjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.8](https://github.com/algolia/instantsearch/compare/react-instantsearch-nextjs@1.2.7...react-instantsearch-nextjs@1.2.8) (2026-06-09)
+
+**Note:** Version bump only for package react-instantsearch-nextjs
+
+
+
+
+
 ## [1.2.7](https://github.com/algolia/instantsearch/compare/react-instantsearch-nextjs@1.2.6...react-instantsearch-nextjs@1.2.7) (2026-06-03)
 
 **Note:** Version bump only for package react-instantsearch-nextjs