+++ packages/react-instantsearch-router-nextjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.35.1](https://github.com/algolia/instantsearch/compare/react-instantsearch-router-nextjs@7.35.0...react-instantsearch-router-nextjs@7.35.1) (2026-06-09)
+
+**Note:** Version bump only for package react-instantsearch-router-nextjs
+
+
+
+
+
 ## [7.35.0](https://github.com/algolia/instantsearch/compare/react-instantsearch-router-nextjs@7.34.1...react-instantsearch-router-nextjs@7.35.0) (2026-06-03)
 
 **Note:** Version bump only for package react-instantsearch-router-nextjs